Fix errors. Add getUserById.

+++ src/main/java/ru/sladkkov/ChatSimbirSoft/repository/MessageRepo.java
@@ -3,5 +3,5 @@ package ru.sladkkov.ChatSimbirSoft.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.sladkkov.ChatSimbirSoft.domain.Message;
 
-public interface MessageRepo extends JpaRepository<Message,Integer> {
+public interface MessageRepo extends JpaRepository<Message,Long> {
 }
