+++ src/main/java/ru/sladkkov/ChatSimbirSoft/repository/RoomRepo.java
@@ -3,5 +3,5 @@ package ru.sladkkov.ChatSimbirSoft.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.sladkkov.ChatSimbirSoft.domain.Room;
 
-public interface RoomRepo extends JpaRepository<Room,Integer> {
+public interface RoomRepo extends JpaRepository<Room,Long> {
 }
